Move adapter mixin tests into a general test file
Instead of essentially repeating the same tests fora ll the adapters,
just test the adapter mixins themselves directly once. The actual
workspace adapters themselves are still tested via the create views.
